softwaredevelopmentproject-master/requirements.txt
Django
djangorestframework
markdown                # Markdown support for the browsable API.
django-filter           # Filtering support
Pillow
stripe
djangorestframework-simplejwt
gunicorn
django-cors-headers
#mysqlclient
mysql-connector-python